pyproject.toml
# SPDX-FileCopyrightText: 2024 Tjark Sievers
# SPDX-FileCopyrightText: 2025 Tjark Sievers
#
# SPDX-License-Identifier: MIT

[project]
name = "quant-met"
version = "0.0.26"
description = "Calculate superconductivity in flat-band systems."
authors = [{name="Tjark Sievers", email="tsievers@physnet.uni-hamburg.de"}]
readme = "README.md"
homepage = "https://quant-met.tjarksievers.de"
repository = "https://github.com/Ruberhauptmann/quant-met"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "h5py>=3.12.1",
    "matplotlib>=3.10.0",
    "numba>=0.60.0",
    "numpy<2.1",
    "numpydantic>=1.6.6",
    "pandas>=2.2.3",
    "pydantic>=2.10.4",
    "pythtb>=1.8.0",
    "pyyaml>=6.0.2",
    "scipy>=1.15.0",
    "tables>=3.10.2",
]

[project.scripts]
quant-met = 'quant_met.cli:cli'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Config for ruff
[tool.ruff]
line-length = 100
force-exclude = true
extend-exclude = ["docs/source/conf.py", "tests/", "edipack2triqs/"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203", # one-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "COM812",
    "ISC001",
    "Q000",
    "Q003",
    "S101", # This checks against using assert
    "PLR0913", # This checks against using more than 5 function arguments
    "PLR0915", # This checks against too long functions
    "PLR0912", # Check against too many branches
    "C901", # Check against too complex functions
    "PLR2004", # Check against "magic values"
    "T201", # Warn about print statements
    "PLW0603" # Warn about global updates
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pre-commit>=4.0.1",
    "rust-just>=1.40.0",
    "scriv>=1.5.1",
]
lint = [
    "ruff>=0.11.6",
]
test = [
    "hypothesis>=6.131.6",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-integration>=0.2.3",
    "pytest-mock>=3.14.0",
    "pytest-regressions>=2.7.0",
]
docs = [
    "autodoc-pydantic>=2.2.0",
    "pydata-sphinx-theme>=0.16.1",
    "sphinx>=8.2.3",
    "sphinx-design>=0.6.1",
    "sphinx-gallery>=0.19.0",
]
type = [
    "mypy>=1.15.0",
    "types-pyyaml>=6.0.12.20250402",
]
